fix pytest import issues and setup more extensive output
