+++ pyproject.toml
@@ -20,3 +20,8 @@ classifiers = [
 [project.urls]
 "Homepage" = "https://github.com/pypa/sampleproject"
 "Bug Tracker" = "https://github.com/pypa/sampleproject/issues"
+
+[tool.pytest.ini_options]
+pythonpath = [
+  ".", "src",
+]
