+++ src/example_package_yeerahdi/example.py
@@ -1,6 +1,2 @@
 def add_one(number):
     return number + 1
-
-
-def square_number(number):
-    return number ** 2
