+++ src/example_package_yeerahdi/example_square.py
@@ -1,0 +1,2 @@
+def square_number(number):
+    return number ** 2
